refactor: one media download hook behind the video and audio names
